Migrated the new setupLogging function
